src/pages/api/v1/company/[companyId]/salary/employee/[employeeId]/index.test.ts
import { NextApiRequest, NextApiResponse } from 'next';
import handler from '@/pages/api/v1/company/[companyId]/salary/employee/[employeeId]/index';

let req: jest.Mocked<NextApiRequest>;
let res: jest.Mocked<NextApiResponse>;

beforeEach(() => {
  req = {
    headers: {},
    query: {},
    method: '',
    json: jest.fn(),
  } as unknown as jest.Mocked<NextApiRequest>;

  res = {
    status: jest.fn().mockReturnThis(),
    json: jest.fn(),
  } as unknown as jest.Mocked<NextApiResponse>;
});

afterEach(() => {
  jest.clearAllMocks();
});
describe('CreateAnSalaryBookkeeping API Handler Tests', () => {
  it('should return error if id is not provided', async () => {
    req.method = 'POST';
    await handler(req, res);
    expect(res.status).toHaveBeenCalledWith(422);
    expect(res.json).toHaveBeenCalledWith({
      powerby: expect.any(String),
      success: expect.any(Boolean),
      code: expect.stringContaining('422'),
      message: expect.any(String),
      payload: expect.any(Object),
    });
  });
  it('should return error if body element is not provided', async () => {
    req.method = 'POST';
    req.query = { employeeId: '3' };
    req.body = {
      start_date: '2023-03-01',
      end_date: '2023-03-31',
    };
    await handler(req, res);
    expect(res.status).toHaveBeenCalledWith(422);
    expect(res.json).toHaveBeenCalledWith({
      powerby: expect.any(String),
      success: expect.any(Boolean),
      code: expect.stringContaining('422'),
      message: expect.any(String),
      payload: expect.any(Object),
    });
  });
  it('should create an employee salary bookkeeping', async () => {
    req.method = 'POST';
    req.query = { employeeId: '3' };
    req.body = {
      start_date: '2023-03-01',
      end_date: '2023-03-31',
      description: 'March salary',
    };
    await handler(req, res);
    expect(res.status).toHaveBeenCalledWith(201);
    expect(res.json).toHaveBeenCalledWith({
      powerby: expect.any(String),
      success: expect.any(Boolean),
      code: expect.stringContaining('201'),
      message: expect.any(String),
      payload: expect.any(Object),
    });
  });
});
